added mailer,uploading file,chatbox

+++ controllers/api/v1/posts_api.js
@@ -24,14 +24,18 @@ module.exports.destroy = async function (req, res) {
     try {
       const post = await Post.findById(req.params.id);
   
-      //   if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
         post.deleteOne();
        
         await Comment.deleteMany({ post: req.params.id });
        
         return res.json(200,{
             message:"post and associated comments deleted successfully"
+        });}else{
+          return res.json(401,{
+            message:"You cannot delete the post"
         });
+        }
   
     } catch (err) {
       console.log('*********',err);
